Large recode and cleanup
Haven't updated in a while. Was busy with this update (along with other
things like assignments and actually enjoying my holidays). Code is now
alot cleaner and easier to work with:

Video code GLES implementation is now seperate from rest of code. This
allows me to easily create a new implementation in the future, and
allows the editor to have its own video code without having to rewrite
GLES code.

Common utilities like logging and object management is now in a seperate
library called libmint. Keeping it seperate from the rest of the code
and allows me to reuse it in different projects.